Dashboard/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.5.3
SQLAlchemy==2.0.25
databases==0.8.0
mysql-connector-python==8.2.1
redis==5.0.1
python-dotenv==1.0.1
openai==1.10.0
langchain==0.1.2
nltk==3.8.1
transformers==4.37.2
torch==2.2.0
google-api-python-client==2.115.0
gTTS==2.5.1
pydub==0.25.1
moviepy==1.0.3
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
gunicorn==21.2.0
python-jose==3.3.0
fpdf==1.7.2
